backend/requirements.txt
# FastAPI 和核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.10.1

# 数据验证和序列化
marshmallow==3.20.1

# 缓存
redis==5.0.1

# 时间和时区
pytz==2023.3
python-dateutil==2.8.2

# HTTP 客户端
httpx==0.25.2
requests==2.31.0

# 日志和监控
python-json-logger==2.0.7

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
isort==5.13.2
